Large reorganization and partial rewrite
This gives almost all types a JSX prefix to more easily
distinguish the types from the various Parser and Lexer types
that already exist in the JDK and in third party packages with
which this package is likely to be used.

The Position type has been removed and has been replaced with
the LexicalPositionTypes from the new io7m/jlexing package.